requirements.txt
adal==1.2.2
azure-common==1.1.24
azure-core==1.2.1
azure-identity==1.2.0
azure-keyvault==4.0.0
azure-keyvault-keys==4.0.0
azure-keyvault-secrets==4.0.0
azure-storage-blob==12.1.0
azure-storage-common==2.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
isodate==0.6.0
msal==1.1.0
msal-extensions==0.1.3
msrest==0.6.10
msrestazure==0.6.2
oauthlib==3.1.0
portalocker==1.5.2
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
